refactor: menu items with href
