+++ components/header/DropdownMenu/DropdownMenuItem/DropdownMenuItem.types.d.ts
@@ -1,4 +1,5 @@
 export interface IDropdownMenuItem {
   label: string;
   icon: JSX.Element;
+  href?: string;
 }
